Move configured_roles to roles.configured
