+++ mathesar_ui/src/api/rpc/collaborators.ts
@@ -1,7 +1,7 @@
 import { rpcMethodTypeContainer } from '@mathesar/packages/json-rpc-client-builder';
 
-import type { RawConfiguredRole } from './configured_roles';
 import type { RawDatabase } from './databases';
+import type { RawConfiguredRole } from './roles';
 
 export interface RawCollaborator {
   id: number;
