+++ mathesar_ui/src/api/rpc/index.ts
@@ -4,7 +4,6 @@ import { buildRpcApi } from '@mathesar/packages/json-rpc-client-builder';
 
 import { collaborators } from './collaborators';
 import { columns } from './columns';
-import { configured_roles } from './configured_roles';
 import { constraints } from './constraints';
 import { database_setup } from './database_setup';
 import { databases } from './databases';
@@ -20,7 +19,6 @@ export const api = buildRpcApi({
   getHeaders: () => ({ 'X-CSRFToken': Cookies.get('csrftoken') }),
   methodTree: {
     collaborators,
-    configured_roles,
     database_setup,
     databases,
     records,
